arch: ws-9-capability-host-wiring (iter 1, approved)
