+++ crates/ironclaw_loop_support/Cargo.toml
@@ -14,14 +14,15 @@ publish = false
 async-trait = "0.1"
 tracing = "0.1"
 tokio = { version = "1", features = ["sync"] }
+ironclaw_host_api = { path = "../ironclaw_host_api", version = "0.1.0" }
+ironclaw_host_runtime = { path = "../ironclaw_host_runtime", version = "0.1.0" }
 ironclaw_skills = { path = "../ironclaw_skills", version = "0.3.0", default-features = false }
 ironclaw_threads = { path = "../ironclaw_threads", version = "0.1.0" }
 ironclaw_turns = { path = "../ironclaw_turns", version = "0.1.0" }
 serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 thiserror = "2"
 
 [dev-dependencies]
-ironclaw_host_api = { path = "../ironclaw_host_api", version = "0.1.0" }
-serde_json = "1"
 tokio = { version = "1", features = ["macros", "rt"] }
 tracing-test = { version = "0.2", features = ["no-env-filter"] }
