+++ crates/ironclaw_reborn/src/loop_driver_host.rs
@@ -1,22 +1,14 @@
 use std::{
-    collections::{HashMap, VecDeque},
     error::Error,
     fmt,
     sync::{Arc, Mutex},
 };
 
 use async_trait::async_trait;
-use ironclaw_host_api::{
-    CapabilityId, CorrelationId, ExecutionContext, ExtensionId, InvocationId, ResourceEstimate,
-    sha256_digest_token,
-};
-use ironclaw_host_runtime::{
-    HostRuntime, HostRuntimeError, IdempotencyKey, RuntimeBlockedReason, RuntimeCapabilityOutcome,
-    RuntimeCapabilityRequest, RuntimeFailureKind,
-};
 use ironclaw_loop_support::{
-    EmptyLoopCapabilityPort, HostManagedModelGateway, HostSkillContextSource,
-    ThreadBackedLoopContextPort, ThreadBackedLoopModelPort, ThreadBackedLoopTranscriptPort,
+    CapabilitySurfaceProfileFilter, CapabilitySurfaceProfileResolver, EmptyLoopCapabilityPort,
+    HostManagedModelGateway, HostSkillContextSource, ThreadBackedLoopContextPort,
+    ThreadBackedLoopModelPort, ThreadBackedLoopTranscriptPort,
 };
 use ironclaw_threads::{SessionThreadService, ThreadScope};
 
@@ -24,29 +16,26 @@ use crate::model_routes::{ModelRouteError, ModelRouteResolver, ModelSlot};
 
 use ironclaw_turns::{
     CheckpointStateStore, GetCheckpointStateRequest, LoopCheckpointStateRef, LoopCheckpointStore,
-    LoopGateRef, LoopResultRef, PutCheckpointStateRequest, PutLoopCheckpointRequest, RunProfileId,
-    TurnCheckpointId, TurnError, TurnStatus,
+    PutCheckpointStateRequest, PutLoopCheckpointRequest, RunProfileId, TurnCheckpointId, TurnError,
+    TurnStatus,
     run_profile::{
         AgentLoopHostError, AgentLoopHostErrorKind, AppendCapabilityResultRef, BeginAssistantDraft,
-        CapabilityBatchInvocation, CapabilityBatchOutcome, CapabilityDenied,
-        CapabilityDeniedReasonKind, CapabilityDescriptorView, CapabilityFailure,
-        CapabilityFailureKind, CapabilityInvocation, CapabilityOutcome, CapabilityResultMessage,
-        ConcurrencyHint, FinalizeAssistantMessage, HostManagedLoopModelPort,
-        HostManagedLoopPromptPort, InMemoryInstructionMaterializationStore,
-        InstructionMaterializationStore, InstructionSafetyContext, LoopCapabilityPort,
-        LoopCheckpointPort, LoopCheckpointRequest, LoopContextBundle, LoopContextPort,
-        LoopContextRequest, LoopHostMilestoneEmitter, LoopHostMilestoneSink, LoopInputBatch,
-        LoopInputCursor, LoopInputPort, LoopModelBudgetAccountant, LoopModelGateway,
-        LoopModelGatewayError, LoopModelGatewayRequest, LoopModelPolicyGuard, LoopModelPort,
-        LoopModelRequest, LoopModelResponse, LoopProcessRef, LoopProgressEvent, LoopProgressPort,
-        LoopPromptBundle, LoopPromptBundleAuthority, LoopPromptBundleRequest, LoopPromptPort,
-        LoopRunContext, LoopRunInfoPort, LoopSafeSummary, LoopTranscriptPort, NoOpBudgetAccountant,
-        NoOpPolicyGuard, ProcessHandleSummary, StageCheckpointPayloadRequest, UpdateAssistantDraft,
+        CapabilityBatchInvocation, CapabilityBatchOutcome, CapabilityInvocation, CapabilityOutcome,
+        FinalizeAssistantMessage, HostManagedLoopModelPort, HostManagedLoopPromptPort,
+        InMemoryInstructionMaterializationStore, InstructionMaterializationStore,
+        InstructionSafetyContext, LoopCapabilityPort, LoopCheckpointPort, LoopCheckpointRequest,
+        LoopContextBundle, LoopContextPort, LoopContextRequest, LoopHostMilestoneEmitter,
+        LoopHostMilestoneSink, LoopInputBatch, LoopInputCursor, LoopInputPort,
+        LoopModelBudgetAccountant, LoopModelGateway, LoopModelGatewayError,
+        LoopModelGatewayRequest, LoopModelPolicyGuard, LoopModelPort, LoopModelRequest,
+        LoopModelResponse, LoopProgressEvent, LoopProgressPort, LoopPromptBundle,
+        LoopPromptBundleAuthority, LoopPromptBundleRequest, LoopPromptPort, LoopRunContext,
+        LoopRunInfoPort, LoopSafeSummary, LoopTranscriptPort, NoOpBudgetAccountant,
+        NoOpPolicyGuard, StageCheckpointPayloadRequest, UpdateAssistantDraft,
         VisibleCapabilityRequest, VisibleCapabilitySurface,
     },
     runner::ClaimedTurnRun,
 };
-use tokio::sync::Notify;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct TextOnlyLoopHostConfig {
@@ -160,793 +149,6 @@ impl LoopCapabilityPort for SurfaceTrackingLoopCapabilityPort {
     }
 }
 
-#[async_trait]
-pub trait LoopCapabilityInputResolver: Send + Sync {
-    async fn resolve_capability_input(
-        &self,
-        run_context: &LoopRunContext,
-        input_ref: &ironclaw_turns::run_profile::CapabilityInputRef,
-    ) -> Result<serde_json::Value, AgentLoopHostError>;
-}
-
-#[async_trait]
-pub trait LoopCapabilityResultWriter: Send + Sync {
-    async fn write_capability_result(
-        &self,
-        run_context: &LoopRunContext,
-        capability_id: &CapabilityId,
-        output: serde_json::Value,
-    ) -> Result<LoopResultRef, AgentLoopHostError>;
-}
-
-#[derive(Clone)]
-struct SurfaceCapabilitySnapshot {
-    provider: ExtensionId,
-    estimate: ResourceEstimate,
-}
-
-#[derive(Clone, Default)]
-struct SurfaceSnapshot {
-    capabilities: HashMap<CapabilityId, SurfaceCapabilitySnapshot>,
-}
-
-const MAX_IN_MEMORY_DISPATCH_RECORDS: usize = 128;
-
-#[derive(Clone)]
-enum DispatchRecord {
-    InFlight {
-        notify: Arc<Notify>,
-    },
-    RuntimeCompleted {
-        requested_capability_id: CapabilityId,
-        outcome: RuntimeCapabilityOutcome,
-    },
-    LoopCompleted(Result<CapabilityOutcome, AgentLoopHostError>),
-}
-
-#[derive(Default)]
-struct DispatchRecordStore {
-    records: HashMap<String, DispatchRecord>,
-    insertion_order: VecDeque<String>,
-}
-
-impl DispatchRecordStore {
-    fn reserve(&mut self, key: &IdempotencyKey) -> Result<DispatchReservation, AgentLoopHostError> {
-        let key_value = key.as_str().to_string();
-        match self.records.get(key.as_str()).cloned() {
-            Some(DispatchRecord::InFlight { notify }) => Ok(DispatchReservation::Wait(notify)),
-            Some(DispatchRecord::RuntimeCompleted {
-                requested_capability_id,
-                outcome,
-            }) => {
-                self.records.insert(
-                    key_value,
-                    DispatchRecord::InFlight {
-                        notify: Arc::new(Notify::new()),
-                    },
-                );
-                Ok(DispatchReservation::RuntimeCompleted {
-                    requested_capability_id,
-                    outcome,
-                })
-            }
-            Some(DispatchRecord::LoopCompleted(result)) => {
-                Ok(DispatchReservation::LoopCompleted(result))
-            }
-            None => {
-                self.evict_completed_until_below_limit()?;
-                self.insertion_order.push_back(key_value.clone());
-                self.records.insert(
-                    key_value,
-                    DispatchRecord::InFlight {
-                        notify: Arc::new(Notify::new()),
-                    },
-                );
-                Ok(DispatchReservation::Reserved)
-            }
-        }
-    }
-
-    fn record(&mut self, key: &IdempotencyKey, record: DispatchRecord) -> Option<Arc<Notify>> {
-        let previous = self.records.insert(key.as_str().to_string(), record);
-        match previous {
-            Some(DispatchRecord::InFlight { notify }) => Some(notify),
-            _ => None,
-        }
-    }
-
-    fn remove(&mut self, key: &IdempotencyKey) -> Option<Arc<Notify>> {
-        let removed = self.records.remove(key.as_str());
-        self.insertion_order
-            .retain(|candidate| candidate != key.as_str());
-        match removed {
-            Some(DispatchRecord::InFlight { notify }) => Some(notify),
-            _ => None,
-        }
-    }
-
-    fn in_flight_matches(&self, key: &IdempotencyKey, notify: &Arc<Notify>) -> bool {
-        matches!(
-            self.records.get(key.as_str()),
-            Some(DispatchRecord::InFlight { notify: current }) if Arc::ptr_eq(current, notify)
-        )
-    }
-
-    fn evict_completed_until_below_limit(&mut self) -> Result<(), AgentLoopHostError> {
-        let mut scanned = 0;
-        let scan_limit = self.insertion_order.len();
-        while self.records.len() >= MAX_IN_MEMORY_DISPATCH_RECORDS && scanned < scan_limit {
-            let Some(candidate) = self.insertion_order.pop_front() else {
-                break;
-            };
-            scanned += 1;
-            match self.records.get(&candidate) {
-                None => {}
-                Some(DispatchRecord::InFlight { .. }) => self.insertion_order.push_back(candidate),
-                Some(DispatchRecord::RuntimeCompleted { .. })
-                | Some(DispatchRecord::LoopCompleted(_)) => {
-                    self.records.remove(&candidate);
-                }
-            }
-        }
-        if self.records.len() >= MAX_IN_MEMORY_DISPATCH_RECORDS {
-            return Err(AgentLoopHostError::new(
-                AgentLoopHostErrorKind::Unavailable,
-                "capability dispatch record store is full",
-            ));
-        }
-        Ok(())
-    }
-}
-
-enum DispatchReservation {
-    Reserved,
-    Wait(Arc<Notify>),
-    RuntimeCompleted {
-        requested_capability_id: CapabilityId,
-        outcome: RuntimeCapabilityOutcome,
-    },
-    LoopCompleted(Result<CapabilityOutcome, AgentLoopHostError>),
-}
-
-pub struct HostRuntimeLoopCapabilityPort {
-    runtime: Arc<dyn HostRuntime>,
-    run_context: LoopRunContext,
-    visible_request: ironclaw_host_runtime::VisibleCapabilityRequest,
-    input_resolver: Arc<dyn LoopCapabilityInputResolver>,
-    result_writer: Arc<dyn LoopCapabilityResultWriter>,
-    milestone_sink: Option<Arc<dyn LoopHostMilestoneSink>>,
-    snapshots: Mutex<HashMap<String, SurfaceSnapshot>>,
-    dispatch_records: Mutex<DispatchRecordStore>,
-}
-
-impl HostRuntimeLoopCapabilityPort {
-    pub fn new(
-        runtime: Arc<dyn HostRuntime>,
-        run_context: LoopRunContext,
-        visible_request: ironclaw_host_runtime::VisibleCapabilityRequest,
-        input_resolver: Arc<dyn LoopCapabilityInputResolver>,
-        result_writer: Arc<dyn LoopCapabilityResultWriter>,
-    ) -> Self {
-        Self {
-            runtime,
-            run_context,
-            visible_request,
-            input_resolver,
-            result_writer,
-            milestone_sink: None,
-            snapshots: Mutex::new(HashMap::new()),
-            dispatch_records: Mutex::new(DispatchRecordStore::default()),
-        }
-    }
-
-    pub fn with_milestone_sink(mut self, sink: Arc<dyn LoopHostMilestoneSink>) -> Self {
-        self.milestone_sink = Some(sink);
-        self
-    }
-
-    fn snapshot_for(
-        &self,
-        version: &ironclaw_turns::run_profile::CapabilitySurfaceVersion,
-    ) -> Result<SurfaceSnapshot, AgentLoopHostError> {
-        let snapshots = self.snapshots.lock().map_err(|_| {
-            AgentLoopHostError::new(
-                AgentLoopHostErrorKind::Unavailable,
-                "capability surface snapshot store is unavailable",
-            )
-        })?;
-        snapshots.get(version.as_str()).cloned().ok_or_else(|| {
-            AgentLoopHostError::new(
-                AgentLoopHostErrorKind::StaleSurface,
-                "capability surface is stale or unknown",
-            )
-        })
-    }
-
-    fn reserve_dispatch(
-        &self,
-        key: &IdempotencyKey,
-    ) -> Result<DispatchReservation, AgentLoopHostError> {
-        self.dispatch_records
-            .lock()
-            .map_err(|_| {
-                AgentLoopHostError::new(
-                    AgentLoopHostErrorKind::Unavailable,
-                    "capability dispatch record store is unavailable",
-                )
-            })?
-            .reserve(key)
-    }
-
-    fn dispatch_in_flight_matches(
-        &self,
-        key: &IdempotencyKey,
-        notify: &Arc<Notify>,
-    ) -> Result<bool, AgentLoopHostError> {
-        self.dispatch_records
-            .lock()
-            .map(|records| records.in_flight_matches(key, notify))
-            .map_err(|_| {
-                AgentLoopHostError::new(
-                    AgentLoopHostErrorKind::Unavailable,
-                    "capability dispatch record store is unavailable",
-                )
-            })
-    }
-
-    fn record_runtime_completed(
-        &self,
-        key: &IdempotencyKey,
-        requested_capability_id: CapabilityId,
-        outcome: RuntimeCapabilityOutcome,
-    ) -> Result<(), AgentLoopHostError> {
-        let notify = self
-            .dispatch_records
-            .lock()
-            .map_err(|_| {
-                AgentLoopHostError::new(
-                    AgentLoopHostErrorKind::Unavailable,
-                    "capability dispatch record store is unavailable",
-                )
-            })?
-            .record(
-                key,
-                DispatchRecord::RuntimeCompleted {
-                    requested_capability_id,
-                    outcome,
-                },
-            );
-        if let Some(notify) = notify {
-            notify.notify_waiters();
-        }
-        Ok(())
-    }
-
-    fn record_loop_completed(
-        &self,
-        key: &IdempotencyKey,
-        result: Result<CapabilityOutcome, AgentLoopHostError>,
-    ) -> Result<(), AgentLoopHostError> {
-        let notify = self
-            .dispatch_records
-            .lock()
-            .map_err(|_| {
-                AgentLoopHostError::new(
-                    AgentLoopHostErrorKind::Unavailable,
-                    "capability dispatch record store is unavailable",
-                )
-            })?
-            .record(key, DispatchRecord::LoopCompleted(result));
-        if let Some(notify) = notify {
-            notify.notify_waiters();
-        }
-        Ok(())
-    }
-
-    fn clear_dispatch(&self, key: &IdempotencyKey) -> Result<(), AgentLoopHostError> {
-        let notify = self
-            .dispatch_records
-            .lock()
-            .map_err(|_| {
-                AgentLoopHostError::new(
-                    AgentLoopHostErrorKind::Unavailable,
-                    "capability dispatch record store is unavailable",
-                )
-            })?
-            .remove(key);
-        if let Some(notify) = notify {
-            notify.notify_waiters();
-        }
-        Ok(())
-    }
-
-    fn validate_visible_request_scope(&self) -> Result<(), AgentLoopHostError> {
-        let context = &self.visible_request.context;
-        context.validate().map_err(|_| {
-            AgentLoopHostError::new(
-                AgentLoopHostErrorKind::InvalidInvocation,
-                "capability execution context is invalid",
-            )
-        })?;
-        if context.tenant_id != self.run_context.scope.tenant_id
-            || context.agent_id != self.run_context.scope.agent_id
-            || context.project_id != self.run_context.scope.project_id
-            || context.thread_id.as_ref() != Some(&self.run_context.thread_id)
-            || context.resource_scope.tenant_id != self.run_context.scope.tenant_id
-            || context.resource_scope.agent_id != self.run_context.scope.agent_id
-            || context.resource_scope.project_id != self.run_context.scope.project_id
-            || context.resource_scope.thread_id.as_ref() != Some(&self.run_context.thread_id)
-        {
-            return Err(AgentLoopHostError::new(
-                AgentLoopHostErrorKind::ScopeMismatch,
-                "capability execution context is not scoped to this loop run",
-            ));
-        }
-        Ok(())
-    }
-
-    async fn finish_runtime_outcome(
-        &self,
-        key: &IdempotencyKey,
-        requested_capability_id: &CapabilityId,
-        outcome: RuntimeCapabilityOutcome,
-    ) -> Result<CapabilityOutcome, AgentLoopHostError> {
-        let result = runtime_outcome_to_loop(
-            &self.run_context,
-            self.result_writer.as_ref(),
-            requested_capability_id,
-            outcome.clone(),
-        )
-        .await;
-        if should_retry_result_write(&outcome, &result) {
-            self.record_runtime_completed(key, requested_capability_id.clone(), outcome)?;
-            return result;
-        }
-        self.record_loop_completed(key, result.clone())?;
-        result
-    }
-
-    async fn wait_for_dispatch_completion(
-        &self,
-        key: &IdempotencyKey,
-        notify: Arc<Notify>,
-    ) -> Result<(), AgentLoopHostError> {
-        let notified = notify.notified();
-        tokio::pin!(notified);
-        if self.dispatch_in_flight_matches(key, &notify)? {
-            notified.await;
-        }
-        Ok(())
-    }
-
-    async fn emit_capability_invoked(
-        &self,
-        capability_id: CapabilityId,
-    ) -> Result<(), AgentLoopHostError> {
-        if let Some(milestone_sink) = &self.milestone_sink {
-            let milestones =
-                LoopHostMilestoneEmitter::new(self.run_context.clone(), Arc::clone(milestone_sink));
-            milestones.capability_invoked(capability_id).await?;
-        }
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl LoopCapabilityPort for HostRuntimeLoopCapabilityPort {
-    async fn visible_capabilities(
-        &self,
-        _request: VisibleCapabilityRequest,
-    ) -> Result<VisibleCapabilitySurface, AgentLoopHostError> {
-        self.validate_visible_request_scope()?;
-        let runtime_surface = self
-            .runtime
-            .visible_capabilities(self.visible_request.clone())
-            .await
-            .map_err(host_runtime_error)?;
-        let version = loop_surface_version(runtime_surface.version.as_str())?;
-        let mut snapshot = SurfaceSnapshot::default();
-        let descriptors = runtime_surface
-            .capabilities
-            .into_iter()
-            .map(|capability| {
-                let capability_id = capability.descriptor.id.clone();
-                snapshot.capabilities.insert(
-                    capability_id.clone(),
-                    SurfaceCapabilitySnapshot {
-                        provider: capability.descriptor.provider.clone(),
-                        estimate: capability.estimated_resources.clone(),
-                    },
-                );
-                CapabilityDescriptorView {
-                    capability_id,
-                    provider: Some(capability.descriptor.provider),
-                    runtime: capability.descriptor.runtime,
-                    safe_name: capability.descriptor.id.as_str().to_string(),
-                    safe_description: capability.descriptor.description,
-                    // WS-9 derives this from the underlying
-                    // `CapabilityDescriptor.effects` Vec. Until that landing,
-                    // default to `Exclusive` (the conservative choice — forces
-                    // serial invocation and never accidentally enables
-                    // parallel side effects).
-                    concurrency_hint: ConcurrencyHint::Exclusive,
-                }
-            })
-            .collect();
-
-        let mut snapshots = self.snapshots.lock().map_err(|_| {
-            AgentLoopHostError::new(
-                AgentLoopHostErrorKind::Unavailable,
-                "capability surface snapshot store is unavailable",
-            )
-        })?;
-        snapshots.clear();
-        snapshots.insert(version.as_str().to_string(), snapshot);
-
-        Ok(VisibleCapabilitySurface {
-            version,
-            descriptors,
-        })
-    }
-
-    async fn invoke_capability(
-        &self,
-        request: CapabilityInvocation,
-    ) -> Result<CapabilityOutcome, AgentLoopHostError> {
-        let snapshot = self.snapshot_for(&request.surface_version)?;
-        let Some(capability) = snapshot.capabilities.get(&request.capability_id).cloned() else {
-            return Ok(CapabilityOutcome::Denied(CapabilityDenied {
-                reason_kind: capability_denied_reason_kind("outside_visible_surface")?,
-                safe_summary: "capability was not visible on the cited surface".to_string(),
-            }));
-        };
-        let Some(trust_decision) = self
-            .visible_request
-            .provider_trust
-            .get(&capability.provider)
-            .cloned()
-        else {
-            return Ok(CapabilityOutcome::Denied(CapabilityDenied {
-                reason_kind: capability_denied_reason_kind("missing_provider_trust")?,
-                safe_summary: "capability provider trust is unavailable".to_string(),
-            }));
-        };
-        let idempotency_key = invocation_idempotency_key(&self.run_context, &request)?;
-        loop {
-            match self.reserve_dispatch(&idempotency_key)? {
-                DispatchReservation::Reserved => break,
-                DispatchReservation::Wait(notify) => {
-                    self.wait_for_dispatch_completion(&idempotency_key, notify)
-                        .await?;
-                }
-                DispatchReservation::RuntimeCompleted {
-                    requested_capability_id,
-                    outcome,
-                } => {
-                    return self
-                        .finish_runtime_outcome(&idempotency_key, &requested_capability_id, outcome)
-                        .await;
-                }
-                DispatchReservation::LoopCompleted(result) => return result,
-            }
-        }
-        let input = match self
-            .input_resolver
-            .resolve_capability_input(&self.run_context, &request.input_ref)
-            .await
-        {
-            Ok(input) => input,
-            Err(error) => {
-                if let Err(clear_error) = self.clear_dispatch(&idempotency_key) {
-                    tracing::warn!(
-                        clear_error = %clear_error,
-                        "failed to clear capability dispatch record after input resolution error"
-                    );
-                    return Err(clear_error);
-                }
-                return Err(error);
-            }
-        };
-        let requested_capability_id = request.capability_id.clone();
-
-        if let Err(error) = self
-            .emit_capability_invoked(request.capability_id.clone())
-            .await
-        {
-            if let Err(clear_error) = self.clear_dispatch(&idempotency_key) {
-                tracing::warn!(
-                    clear_error = %clear_error,
-                    "failed to clear capability dispatch record after milestone emission error"
-                );
-                return Err(clear_error);
-            }
-            return Err(error);
-        }
-        let outcome = match self
-            .runtime
-            .invoke_capability(
-                RuntimeCapabilityRequest::new(
-                    invocation_context_from_visible(&self.visible_request.context),
-                    request.capability_id,
-                    capability.estimate,
-                    input,
-                    trust_decision,
-                )
-                .with_idempotency_key(idempotency_key.clone()),
-            )
-            .await
-        {
-            Ok(outcome) => outcome,
-            Err(error) => {
-                if let Err(clear_error) = self.clear_dispatch(&idempotency_key) {
-                    tracing::warn!(
-                        clear_error = %clear_error,
-                        "failed to clear capability dispatch record after host runtime error"
-                    );
-                    return Err(clear_error);
-                }
-                return Err(host_runtime_error(error));
-            }
-        };
-        self.finish_runtime_outcome(&idempotency_key, &requested_capability_id, outcome)
-            .await
-    }
-
-    async fn invoke_capability_batch(
-        &self,
-        request: CapabilityBatchInvocation,
-    ) -> Result<CapabilityBatchOutcome, AgentLoopHostError> {
-        let mut outcomes = Vec::new();
-        let mut stopped_on_suspension = false;
-        for invocation in request.invocations {
-            let outcome = self.invoke_capability(invocation).await?;
-            let is_suspension = outcome.is_suspension();
-            outcomes.push(outcome);
-            if request.stop_on_first_suspension && is_suspension {
-                stopped_on_suspension = true;
-                break;
-            }
-        }
-        Ok(CapabilityBatchOutcome {
-            outcomes,
-            stopped_on_suspension,
-        })
-    }
-}
-
-fn should_retry_result_write(
-    outcome: &RuntimeCapabilityOutcome,
-    result: &Result<CapabilityOutcome, AgentLoopHostError>,
-) -> bool {
-    matches!(outcome, RuntimeCapabilityOutcome::Completed(_))
-        && matches!(
-            result,
-            Err(error)
-                if matches!(
-                    error.kind,
-                    AgentLoopHostErrorKind::Unavailable
-                        | AgentLoopHostErrorKind::TranscriptWriteFailed
-                )
-        )
-}
-
-fn invocation_context_from_visible(base: &ExecutionContext) -> ExecutionContext {
-    // The visible-capability request is the host-vetted authority envelope for
-    // this loop run. Each dispatch gets fresh invocation/correlation ids and no
-    // inherited process parentage, while identity, trust, grants, mounts, and
-    // run scope remain bound to the visible surface that the model cited.
-    let mut context = base.clone();
-    let invocation_id = InvocationId::new();
-    context.invocation_id = invocation_id;
-    context.correlation_id = CorrelationId::new();
-    context.process_id = None;
-    context.parent_process_id = None;
-    context.resource_scope.invocation_id = invocation_id;
-    context
-}
-
-fn invocation_idempotency_key(
-    run_context: &LoopRunContext,
-    request: &CapabilityInvocation,
-) -> Result<IdempotencyKey, AgentLoopHostError> {
-    let payload = format!(
-        "loop-capability\nrun={}\nsurface={}\ncapability={}\ninput={}",
-        run_context.run_id,
-        request.surface_version.as_str(),
-        request.capability_id.as_str(),
-        request.input_ref.as_str()
-    );
-    IdempotencyKey::new(format!(
-        "loop-capability:{}",
-        sha256_digest_token(payload.as_bytes())
-    ))
-    .map_err(host_runtime_error)
-}
-
-fn loop_surface_version(
-    version: &str,
-) -> Result<ironclaw_turns::run_profile::CapabilitySurfaceVersion, AgentLoopHostError> {
-    ironclaw_turns::run_profile::CapabilitySurfaceVersion::new(version).map_err(|_| {
-        AgentLoopHostError::new(
-            AgentLoopHostErrorKind::Internal,
-            "host runtime capability surface version could not be represented",
-        )
-    })
-}
-
-async fn runtime_outcome_to_loop(
-    run_context: &LoopRunContext,
-    result_writer: &(dyn LoopCapabilityResultWriter + Send + Sync),
-    requested_capability_id: &CapabilityId,
-    outcome: RuntimeCapabilityOutcome,
-) -> Result<CapabilityOutcome, AgentLoopHostError> {
-    ensure_runtime_outcome_matches(requested_capability_id, &outcome)?;
-    Ok(match outcome {
-        RuntimeCapabilityOutcome::Completed(completed) => {
-            let result_ref = result_writer
-                .write_capability_result(
-                    run_context,
-                    &completed.capability_id,
-                    completed.output.clone(),
-                )
-                .await?;
-            CapabilityOutcome::Completed(CapabilityResultMessage {
-                result_ref,
-                safe_summary: "capability completed".to_string(),
-                terminate_hint: false,
-            })
-        }
-        RuntimeCapabilityOutcome::ApprovalRequired(gate) => CapabilityOutcome::ApprovalRequired {
-            gate_ref: loop_gate_ref("approval", gate.approval_request_id.to_string())?,
-            safe_summary: blocked_summary(gate.reason).to_string(),
-        },
-        RuntimeCapabilityOutcome::AuthRequired(gate) => CapabilityOutcome::AuthRequired {
-            gate_ref: loop_gate_ref("auth", gate.gate_id.to_string())?,
-            safe_summary: blocked_summary(gate.reason).to_string(),
-        },
-        RuntimeCapabilityOutcome::ResourceBlocked(gate) => CapabilityOutcome::ResourceBlocked {
-            gate_ref: loop_gate_ref("resource", gate.gate_id.to_string())?,
-            safe_summary: blocked_summary(gate.reason).to_string(),
-        },
-        RuntimeCapabilityOutcome::SpawnedProcess(process) => {
-            CapabilityOutcome::SpawnedProcess(ProcessHandleSummary {
-                process_ref: LoopProcessRef::new(format!("process:{}", process.process_id))
-                    .map_err(|_| {
-                        AgentLoopHostError::new(
-                            AgentLoopHostErrorKind::Internal,
-                            "process ref could not be represented",
-                        )
-                    })?,
-                safe_summary: "capability spawned background work".to_string(),
-            })
-        }
-        RuntimeCapabilityOutcome::Failed(failure) => {
-            if failure.kind == RuntimeFailureKind::Authorization {
-                CapabilityOutcome::Denied(CapabilityDenied {
-                    reason_kind: capability_denied_reason_kind(failure.kind.as_str())?,
-                    safe_summary: runtime_safe_summary(
-                        failure.message,
-                        "capability authorization denied",
-                    ),
-                })
-            } else {
-                CapabilityOutcome::Failed(CapabilityFailure {
-                    error_kind: runtime_failure_kind_to_loop(failure.kind),
-                    safe_summary: runtime_safe_summary(
-                        failure.message,
-                        "capability invocation failed",
-                    ),
-                })
-            }
-        }
-        RuntimeCapabilityOutcome::Unknown(unknown) => {
-            CapabilityOutcome::Failed(CapabilityFailure {
-                error_kind: CapabilityFailureKind::unknown(unknown.kind).map_err(|_| {
-                    AgentLoopHostError::new(
-                        AgentLoopHostErrorKind::Internal,
-                        "unknown capability outcome kind could not be represented",
-                    )
-                })?,
-                safe_summary: runtime_safe_summary(
-                    unknown.message,
-                    "capability invocation returned an unknown outcome",
-                ),
-            })
-        }
-    })
-}
-
-fn runtime_failure_kind_to_loop(kind: RuntimeFailureKind) -> CapabilityFailureKind {
-    match kind {
-        RuntimeFailureKind::Authorization => CapabilityFailureKind::Authorization,
-        RuntimeFailureKind::Backend => CapabilityFailureKind::Backend,
-        RuntimeFailureKind::Cancelled => CapabilityFailureKind::Cancelled,
-        RuntimeFailureKind::Dispatcher => CapabilityFailureKind::Dispatcher,
-        RuntimeFailureKind::InvalidInput => CapabilityFailureKind::InvalidInput,
-        RuntimeFailureKind::MissingRuntime => CapabilityFailureKind::MissingRuntime,
-        RuntimeFailureKind::Network => CapabilityFailureKind::Network,
-        RuntimeFailureKind::OutputTooLarge => CapabilityFailureKind::OutputTooLarge,
-        RuntimeFailureKind::Process => CapabilityFailureKind::Process,
-        RuntimeFailureKind::Resource => CapabilityFailureKind::Resource,
-        RuntimeFailureKind::Unknown => CapabilityFailureKind::Permanent,
-        _ => CapabilityFailureKind::Permanent,
-    }
-}
-
-fn ensure_runtime_outcome_matches(
-    expected: &CapabilityId,
-    outcome: &RuntimeCapabilityOutcome,
-) -> Result<(), AgentLoopHostError> {
-    let actual = match outcome {
-        RuntimeCapabilityOutcome::Completed(completed) => &completed.capability_id,
-        RuntimeCapabilityOutcome::ApprovalRequired(gate) => &gate.capability_id,
-        RuntimeCapabilityOutcome::AuthRequired(gate) => &gate.capability_id,
-        RuntimeCapabilityOutcome::ResourceBlocked(gate) => &gate.capability_id,
-        RuntimeCapabilityOutcome::SpawnedProcess(process) => &process.capability_id,
-        RuntimeCapabilityOutcome::Failed(failure) => &failure.capability_id,
-        RuntimeCapabilityOutcome::Unknown(unknown) => &unknown.capability_id,
-    };
-    if actual != expected {
-        return Err(AgentLoopHostError::new(
-            AgentLoopHostErrorKind::Internal,
-            "host runtime returned outcome for a different capability",
-        ));
-    }
-    Ok(())
-}
-
-fn capability_denied_reason_kind(
-    value: impl Into<String>,
-) -> Result<CapabilityDeniedReasonKind, AgentLoopHostError> {
-    CapabilityDeniedReasonKind::unknown(value).map_err(|_| {
-        AgentLoopHostError::new(
-            AgentLoopHostErrorKind::Internal,
-            "capability denied reason kind could not be represented",
-        )
-    })
-}
-
-fn runtime_safe_summary(message: Option<String>, fallback: &'static str) -> String {
-    message
-        .and_then(|summary| LoopSafeSummary::new(summary).ok())
-        .map(|summary| summary.to_string())
-        .unwrap_or_else(|| fallback.to_string())
-}
-
-fn loop_gate_ref(kind: &str, id: String) -> Result<LoopGateRef, AgentLoopHostError> {
-    LoopGateRef::new(format!("gate:{kind}-{id}")).map_err(|_| {
-        AgentLoopHostError::new(
-            AgentLoopHostErrorKind::Internal,
-            "capability gate ref could not be represented",
-        )
-    })
-}
-
-fn blocked_summary(reason: RuntimeBlockedReason) -> &'static str {
-    match reason {
-        RuntimeBlockedReason::ApprovalRequired => "capability requires approval",
-        RuntimeBlockedReason::AuthRequired => "capability requires authentication",
-        RuntimeBlockedReason::ResourceLimit => "capability is blocked by resource limits",
-        RuntimeBlockedReason::ResourceUnavailable => "capability resources are unavailable",
-    }
-}
-
-fn host_runtime_error(error: HostRuntimeError) -> AgentLoopHostError {
-    match error {
-        HostRuntimeError::InvalidRequest { reason } => AgentLoopHostError::new(
-            AgentLoopHostErrorKind::InvalidInvocation,
-            runtime_safe_summary(Some(reason), "host runtime rejected capability request"),
-        ),
-        HostRuntimeError::Unavailable { reason } => AgentLoopHostError::new(
-            AgentLoopHostErrorKind::Unavailable,
-            runtime_safe_summary(
-                Some(reason),
-                "host runtime capability service is unavailable",
-            ),
-        ),
-    }
-}
-
 pub struct RebornLoopDriverHostFactory<S, G>
 where
     S: SessionThreadService + ?Sized,
@@ -1033,6 +235,28 @@ where
         request: RebornLoopDriverHostRequest,
     ) -> Result<RebornLoopDriverHost, RebornLoopDriverHostError> {
         self.build_text_only_host_with_capabilities(request, Arc::new(EmptyLoopCapabilityPort))
+            .await
+    }
+
+    pub async fn build_text_only_host_with_profiled_capabilities(
+        &self,
+        request: RebornLoopDriverHostRequest,
+        capabilities: Arc<dyn LoopCapabilityPort>,
+        surface_resolver: Arc<dyn CapabilitySurfaceProfileResolver>,
+    ) -> Result<RebornLoopDriverHost, RebornLoopDriverHostError> {
+        validate_claimed_run_context(&request.claimed_run, &request.loop_run_context)?;
+        validate_thread_scope(&self.thread_scope, &request.loop_run_context)?;
+        let allow_set = Arc::new(
+            surface_resolver
+                .resolve(&request.loop_run_context)
+                .await
+                .map_err(|error| RebornLoopDriverHostError::InvalidRequest {
+                    reason: format!("capability surface profile resolver failed: {error}"),
+                })?,
+        );
+        let capabilities: Arc<dyn LoopCapabilityPort> =
+            Arc::new(CapabilitySurfaceProfileFilter::new(capabilities, allow_set));
+        self.build_text_only_host_with_capabilities(request, capabilities)
             .await
     }
